Upgrade from Yarn Classic to Yarn Modern
Before this change, the instructions in the README were wrong. The
instructions in the README told users that they had to have Yarn
installed, and they linked to the Web site for Yarn Modern.
Unfortunately, package.json said this:

  "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"

That Yarn version 1.22.22 is a version of Yarn Classic, not a version of
Yarn Modern. If you use Yarn Modern (like the README says to do), then
it will automatically make a bunch of changes that are unrelated to
whatever you’re doing.

In order to fix this issue, this commit updates the repo to use Yarn
Modern instead of Yarn Classic. We probably should have been using Yarn
Modern to begin with.